+++ Editor/Utils/DisableUnityLogScoop.cs
@@ -1,0 +1,17 @@
+using System;
+
+namespace SaintsBuild.Editor.Utils
+{
+    public class DisableUnityLogScoop: IDisposable
+    {
+        public DisableUnityLogScoop()
+        {
+            UnityEngine.Debug.unityLogger.logEnabled = false;
+        }
+
+        public void Dispose()
+        {
+            UnityEngine.Debug.unityLogger.logEnabled = true;
+        }
+    }
+}