+++ Samples~/Scripts/SubContent.cs
@@ -1,21 +1,15 @@
 using UnityEngine;
 
-namespace SaintsBuild.Samples
+namespace SaintsBuild.Samples.Scripts
 {
     public class SubContent : MonoBehaviour, IPostProcess
     {
 
 #if UNITY_EDITOR
-        public void EditorOnPostProcess(PostProcessInfo postProcessInfo)
+        public bool EditorOnPostProcess(PostProcessInfo postProcessInfo)
         {
-            if (postProcessInfo.PrefabDangerousDestroy())  // hide it so we don't destroy the prefab and save it
-            {
-                gameObject.SetActive(false);
-            }
-            else  // can safely destroy
-            {
-                DestroyImmediate(gameObject);
-            }
+            DestroyImmediate(gameObject, true);
+            return true;
         }
 #endif
     }